[components][i2c] 软件 I2C 分层
