+++ components/drivers/i2c/SConscript
@@ -10,6 +10,7 @@ i2c_dev.c
 if GetDepend('RT_USING_I2C_BITOPS'):
     src = src + ['i2c-bit-ops.c']
 if GetDepend('RT_USING_SOFT_I2C'):
+    src = src + ['soft_i2c_ops.c']
     src = src + ['soft_i2c.c']
 
 # The set of source files associated with this SConscript file.
